refactor: rename PreferencesStep to DataUsageStep
