+++ frontend/src/metabase/setup/components/DataUsageStep/index.ts
@@ -1,0 +1,1 @@
+export * from "./DataUsageStep";
