--- frontend/src/metabase/setup/components/PreferencesStep/index.ts
@@ -1,1 +1,0 @@
-export * from "./PreferencesStep";
